requirement.txt
# Core Django Framework
django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# CORS handling for API
django-cors-headers==4.3.1

# Environment variable management
django-environ==0.11.2

# API Documentation (Swagger)
drf-yasg==1.21.7

# Background task processing
celery==5.3.4
redis==5.0.1

# Message broker for Celery
amqp==5.2.0
kombu==5.3.4

# Additional useful packages
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0